requirements.txt
aiofiles==24.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.12.14
aiosignal==1.4.0
alembic==1.16.4
annotated-types==0.7.0
anthropic==0.60.0
anyio==4.9.0
appdirs==1.4.4
appnope==0.1.4
asgiref==3.9.1
asttokens==3.0.0
asyncpg==0.30.0
attrs==25.3.0
backoff==2.2.1
bcrypt==4.3.0
beautifulsoup4==4.12.3
blinker==1.9.0
boolean.py==5.0
Brotli==1.1.0
browserbase==1.4.0
build==1.2.2.post1
cachetools==5.5.2
certifi==2025.7.14
cffi==1.17.1
charset-normalizer==3.4.2
chroma-hnswlib==0.7.6
chromadb==0.5.23
click==8.2.1
cohere==5.16.1
coloredlogs==15.0.1
comm==0.2.2
crewai==0.150.0
crewai-tools==0.58.0
cryptography==45.0.5
dataclasses-json==0.6.7
debugpy==1.8.15
decorator==5.2.1
deprecation==2.1.0
diskcache==5.6.3
distro==1.9.0
dnspython==2.7.0
docker==7.1.0
docstring_parser==0.17.0
durationpy==0.10
ecdsa==0.19.1
elevenlabs==2.8.1
email_validator==2.2.0
embedchain==0.1.128
et_xmlfile==2.0.0
executing==2.2.0
fastapi==0.109.2
fastavro==1.11.1
filelock==3.18.0
flatbuffers==25.2.10
flexcache==0.3
flexparser==0.4
frozenlist==1.7.0
fsspec==2025.7.0
fuzzywuzzy==0.18.0
google-api-core==2.25.1
google-auth==2.29.0
google-auth-oauthlib==1.2.0
google-cloud-core==2.4.3
google-cloud-storage==2.10.0
google-crc32c==1.7.1
google-resumable-media==2.7.2
googleapis-common-protos==1.70.0
gptcache==0.1.44
graphviz==0.21
greenlet==3.0.3
grpcio==1.74.0
h11==0.16.0
h2==4.2.0
harfile==0.3.0
hf-xet==1.1.5
hpack==4.1.0
html5lib==1.1
httpcore==1.0.9
httptools==0.6.4
httpx==0.28.1
httpx-sse==0.4.0
huggingface-hub==0.34.2
humanfriendly==10.0
hyperframe==6.1.0
hypothesis==6.136.3
idna==3.10
importlib_metadata==8.7.0
importlib_resources==6.5.2
ingredient_parser_nlp==2.1.1
iniconfig==2.1.0
instructor==1.10.0
ipykernel==6.30.0
ipython==9.4.0
ipython_pygments_lexers==1.1.1
itsdangerous==2.2.0
jedi==0.19.2
Jinja2==3.1.6
jiter==0.10.0
joblib==1.5.1
json5==0.12.0
json_repair==0.25.2
jsonpatch==1.33
jsonpickle==4.1.1
jsonpointer==3.0.0
jsonref==1.1.0
jsonschema==4.25.0
jsonschema-specifications==2025.4.1
junit-xml==1.9
jupyter_client==8.6.3
jupyter_core==5.8.1
kubernetes==33.1.0
lancedb==0.24.2
langchain==0.3.27
langchain-cohere==0.3.5
langchain-community==0.3.27
langchain-core==0.3.72
langchain-experimental==0.3.4
langchain-openai==0.2.14
langchain-text-splitters==0.3.9
langsmith==0.3.45
Levenshtein==0.27.1
license-expression==30.4.4
litellm==1.74.3
Mako==1.3.10
markdown-it-py==3.0.0
MarkupSafe==3.0.2
marshmallow==3.26.1
matplotlib-inline==0.1.7
mcp==1.12.2
mdurl==0.1.2
mem0ai==0.1.115
mmh3==5.1.0
monotonic==1.6
mpmath==1.3.0
msgpack==1.1.1
multidict==6.6.3
mypy_extensions==1.1.0
nest-asyncio==1.6.0
networkx==3.5
nltk==3.9.1
nodeenv==1.9.1
numpy==1.26.4
oauthlib==3.3.1
onnxruntime==1.22.0
openai==1.97.0
openpyxl==3.1.5
opentelemetry-api==1.35.0
opentelemetry-exporter-otlp-proto-common==1.35.0
opentelemetry-exporter-otlp-proto-grpc==1.35.0
opentelemetry-exporter-otlp-proto-http==1.35.0
opentelemetry-instrumentation==0.56b0
opentelemetry-instrumentation-asgi==0.56b0
opentelemetry-instrumentation-fastapi==0.56b0
opentelemetry-proto==1.35.0
opentelemetry-sdk==1.35.0
opentelemetry-semantic-conventions==0.56b0
opentelemetry-util-http==0.56b0
orjson==3.11.1
overrides==7.7.0
packageurl-python==0.17.1
packaging==23.2
pandas==2.3.1
parso==0.8.4
passlib==1.7.4
pathspec==0.12.1
pdfminer.six==20250506
pdfplumber==0.11.7
pexpect==4.9.0
pillow==11.3.0
Pint==0.24.4
pip-api==0.0.34
platformdirs==4.3.8
playwright==1.44.0
pluggy==1.6.0
portalocker==2.7.0
posthog==3.25.0
prompt_toolkit==3.0.51
propcache==0.3.2
proto-plus==1.26.1
protobuf==6.31.1
psutil==7.0.0
psycopg2-binary==2.9.10
ptyprocess==0.7.0
pure_eval==0.2.3
pyarrow==21.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
pydantic==2.11.7
pydantic-settings==2.10.1
pydantic_core==2.33.2
pydot==4.0.1
pyee==11.1.0
Pygments==2.19.2
PyJWT==2.10.1
pyparsing==3.2.3
pypdf==5.9.0
pypdfium2==4.30.0
PyPika==0.48.9
pyproject_hooks==1.2.0
pyrate-limiter==3.8.1
pyright==1.1.403
pysbd==0.3.4
pytest==8.4.1
pytest-asyncio==1.1.0
pytest-mock==3.14.1
pytest-vcr==1.0.2
python-crfsuite==0.9.11
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-jose==3.3.0
python-Levenshtein==0.27.1
python-multipart==0.0.20
pytube==15.0.0
pytz==2025.2
pyvis==0.3.2
PyYAML==6.0.2
pyzmq==27.0.0
qdrant-client==1.15.0
RapidFuzz==3.13.0
redis==6.2.0
referencing==0.36.2
regex==2024.11.6
requests==2.32.4
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rfc3339-validator==0.1.4
rfc3987==1.3.8
rich==13.9.4
rpds-py==0.26.0
rsa==4.9.1
ruff==0.4.4
black==24.8.0
isort==5.13.2
flake8==7.1.1
flake8-annotations==3.1.1
flake8-docstrings==1.7.0
flake8-fastapi==0.7.0
mypy==1.13.0
sentry-sdk[asgi]==2.22.0
structlog==25.1.0
opentelemetry-instrumentation-fastapi==0.56b0
opentelemetry-exporter-jaeger==1.35.0
prometheus-client==0.23.0
starlette-exporter==0.26.0
loguru==0.7.3
schema==0.7.7
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
sortedcontainers==2.4.0
sounddevice==0.5.1
soundfile==0.13.1
soupsieve==2.7
SQLAlchemy==2.0.41
sqlmodel==0.0.24
sse-starlette==3.0.2
stack-data==0.6.3
stagehand==0.5.0
starlette==0.36.3
sympy==1.14.0
tabulate==0.9.0
tenacity==8.5.0
tiktoken==0.9.0
tokenizers==0.20.3
toml==0.10.2
tomli==2.2.1
tomli_w==1.2.0
tornado==6.5.1
tqdm==4.67.1
traitlets==5.14.3
typer==0.16.0
types-python-dateutil==2.9.0.20250708
types-requests==2.32.4.20250611
typing-inspect==0.9.0
typing-inspection==0.4.1
typing_extensions==4.14.1
tzdata==2025.2
uri-template==1.3.0
urllib3==2.5.0
uv==0.8.3
uvicorn==0.27.1
uvloop==0.21.0
vcrpy==7.0.0
watchfiles==1.1.0
wcwidth==0.2.13
webcolors==24.11.1
webencodings==0.5.1
websocket-client==1.8.0
websockets==15.0.1
wrapt==1.17.2
yarl==1.20.1
zipp==3.23.0
zstandard==0.23.0
